create immutable and mutable versions of text index
