+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -1,5 +1,7 @@
+mod immutable_text_index;
 mod inverted_index;
 mod mmap_inverted_index;
+mod mutable_text_index;
 mod posting_list;
 mod postings_iterator;
 pub mod text_index;
